Browser re-exports from Node, only RN has cast
