+++ sdk/tables/data-tables/src/getSecondaryUrlFromPrimary-react-native.mts
@@ -1,4 +1,18 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-export { getSecondaryUrlFromPrimary } from "./getSecondaryUrlFromPrimary-browser.mjs";
+const SecondaryLocationAccountSuffix = "-secondary";
+
+/**
+ * Utility function that calculates the secondary URL for a table instance given the primary URL.
+ */
+export function getSecondaryUrlFromPrimary(primaryUrl: string): string {
+  const parsedPrimaryUrl = new URL(primaryUrl) as { hostname: string; toString(): string };
+  const host = parsedPrimaryUrl.hostname.split(".");
+  if (host.length > 1) {
+    host[0] = `${host[0]}${SecondaryLocationAccountSuffix}`;
+  }
+  parsedPrimaryUrl.hostname = host.join(".");
+
+  return parsedPrimaryUrl.toString();
+}
